Update folder structure - move component files in root path
