+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import StyledComponentsRegistry from "./components/styledRegistry";
+import StyledComponentsRegistry from "../components/styledRegistry";
 
 const inter = Inter({ subsets: ["latin"] });
