+++ components/styledRegistry/index.tsx
